Add data.onRenderEnd field to render() and renderSync()
To fix #1166, We must render tag plugins before 'post_render' filter works.

I added data.onRenderEnd field to render() and renderSync().
This field is a function that receives rendered content and returns processed content.

Now tag plugins be rendered in this function.

Please use this as material for further discussion.