Fix backtick_code_block for ignore tab character.
<raw>``` (tab)
Hello world!
</new>

The expressions like this causes a bug.

In this case, `args` in backtickCodeBlock() is not blank (but a tab character).
Then the `args` will not match both of rAllOptions and rLangCaption.
So `match` will be `null`. and we can't get `match[1]` in this case.

To fix it, I check whether `match` is null or not.