Support root path configure for img tag
Current img tag doesn't prepend root path if it's not a slash. With
this fix, the root path will be prepend to the img src if it's not a
slash.
This fix only works for the hexo tags, does not work for the markdown
syntax of links and images. The issues of markdown links and images should
be fixed in the hexo-renderer-marked module.

Resolves: #1440