Remove unneccessary refs to type attr in helper output snippets.
See Issue #1695: https://github.com/hexojs/hexo/issues/1695

Also see the HTML5 Spec section on the type attr:

http://dev.w3.org/html5/spec-preview/Overview.html#attr-link-type

In short, the use of type tags is purely advisory, and the
spec instructs browsers to infer the resource type before fetching
them. In cases where the inference does not match the declaration,
the resource will not be loaded. When the declaration is left off,
the resource will load. It appears generally safer to leave the tags
off entirely.