Fix the unit test on permalink_root_prefix - virtual
When the site is put in a subdirectory we have the
`url = 'http://yoursite.com/root';` and
`root = '/root/';`
But the test only set `hexo.config.root` but not `hexo.config.url`.
So the test cannot catch the bug in issue #1812.

I cannot pass the unit test when I fix the bug.