Fix the wrong share link when the site is in subdirectory
Fix the bug mentioned in #1812

When the site is in subdirectory, we are suggest to have set url
as 'http://yoursite.com/child' and root as '/child/'.
But the share link in article footer will become
'http://yoursite.com/child/child/slug'.

Fix the bug by replace '/child/' by '/' in virtual link.