Disable to insert extra new line character into the end of backtick code block (fix #3767)
A paragraph follows a backtick code block on a blockquote should be included same blcokquote block.
But the code block always terminates blockquote.
So the blockquote block splits two blocks.

This is caused by the implementation of `backtick_code_block.js`.
It inserts extra new line character into the end of backtick code block.

This patch disables this behavior.